Parser

+++ cube.h
@@ -20,8 +20,8 @@
 #include <stdlib.h>
 #include <fcntl.h>
 #include <string.h>
-#include <mlx.h>
 #include <math.h>
+
 #define WALL '1'
 #define EMPTY '0'
 #define SPACE ' '
@@ -46,9 +46,6 @@
 #define LEFT_ROTATE  65361
 #define PI 3.14
 
-
-
-
 typedef struct s_mlx_data
 {
     void    *mlx;
@@ -72,9 +69,11 @@ typedef struct  s_player
 
 typedef struct s_position
 {
-    int x;
-    int y;
+    float x;
+    float y;
+    //rotation
     int turn_x;
+    //walk
     int turn_y;
     float rotation_angle;
     float rotation_speed;
@@ -97,7 +96,8 @@ typedef struct  s_map
     char        **map;
     t_args      **args;
     t_position  player_pos;
-	t_data     *data;   
+	t_data     *data;
+    
 } t_map;
 
 char	*ft_strjoin(char const *s1, char const *s2);
